Speculative PrintAsObjC crasher fix
Before printing the top-level decls in the generated header, PrintAsObjC sorts them to ensure they will be printed in a stable order. This sort starts by evaluating several criteria which ought to distinguish any ValueDecls, then includes extra criteria to sort ExtensionDecls on the same type.

Apple has received a number of automated crash reports on this extension-only part of the sort comparison. Unfortunately, nobody has filed a bug report about this crash and we have not been able to devise a reproducer, but the crash reports are consistent with the comparison misinterpreting other declaration kinds as ExtensionDecls. In theory, the code above this point ought to always return when it isn't processing two ExtensionDecls, but in practice it seems like it doesn't.

Without a reproducer for this bug, we cannot fully fix it, but we can at least make the code more defensive by adding an explicit check that it is only handling `ExtensionDecl`s. This will cause declarations to be printed in the header in a nondeterministic order, so we will emit a warning soliciting a bug report to gather more information about this problem.

Hopefully fixes rdar://81811616, at least to the point of no longer crashing. No new tests because we don't know how to cause this crash.