[Tests] Fix round-trip test to build with target libraries.
The round-trip test was relying on a host library, but that only works where
we aren't cross-compiling.  Using a target library instead was a little more
complicated than it looked because libswiftDemangling doesn't get built for
the target.  However, it *is* part of the other Swift libraries on the host,
but there's an extra namespace shoehorned in.