Remove an extra load that was generated for coro shadow copies.
This commit is similar to cbb89c78d, but for explosions.  emitShadowCopyIfNeeded()
has some extra code that was added when function arguments were moved out of the
async context to ensure that they are being lifetime-extended, and there is also
code that generates an incorrect load from the shadow copy. However,
emitShadowCopyIfNeeded is supposed return either an alloca or the value, and
IRGenDebugInfo knows to describe the value in the alloca already. The load is
counterproductive it's only valid until whatever register it ends up in is
clobbered, whereas the alloca is valid throughout the function.

This patch removes the load and updates the tests accordingly.

rdar://81565869