[IRGen] Give decl ref to LinkEntity::forAFP.
Previously, a LinkEntity for an AST async function pointer was built by
passing an AbstractFunctionDecl.  Later, decl was used to construct a
SILDeclRef.

That arrangement meant that clients could not construct such a
LinkEntity whose SILDeclRef::Kind could not be inferred from the dynamic
type of the decl from which the SILDeclRef was constructed.  In
particular, clients could not construct a LinkEntity for the initializer
corresponding to a ConstructorDecl.

Here, the arrangment is changed so that the LinkEntity for an AST async
function pointer is built by passing a SILDeclRef.