[TBD] Pass decl ref when adding AFP symbols.
Previously, adding a symbol for an async function pointer to the TBD
entailed passing an AbstractFunctionDecl to
TBDGenVisitor::addAsyncFunctionPointerSymbol.

That arrangment limited clients from specifying a different
SILDeclRef::Kind than the one which could be inferred from the type of
the decl which was passed to the visitor's function.  In particular, it
prevented clients from specifying that the symbol for an initializer
corresponding to a ConstructorDecl should be added to the TBD.

Here, the arrangement is changed so that the visitor's function takes a
SILDeclRef.