[TBD] Add initializers AFPs to TBD.
Previously, the async function pointers for initializers were omitted
from TBD files.  That resulted in a mismatch between the symbols
exported from the IR module and those listed in the TBDS.

Here, that is corrected.  As with regular function pointers, the symbols
for the async function pointers for initializers need to be manually
added just as the symbols for initializers themselves need to be
manually added.

rdar://82045176