Deserializer: call the deserialization callback when a new function declaration is created
This makes sure that all analysis will get the `notifyAddedOrModifiedFunction` notifications.
This fixes a crash in CallerAnalysis, which relies that it's notified for all newly created functions.

This is related to following bug reports for cross-module-optimization:

https://bugs.swift.org/browse/SR-15048
rdar://81701218

Although I believe that there are more bugs involved which need to be fixed.

Unfortunately I don't have a test case for this fix.