[NFC-ish] Create expressions in imported constants lazily
Previously, when ClangImporter imported a macro, it would immediately construct the Swift expression that went in its body. This could be a problem when building the standard library because macro importing could potentially happen during extension binding, before the compiler had processed all of the standard library types enough to answer certain queries about the types involved in the literal. This caused crashes when trying to build the standard library.

This commit delays the synthesis of these expressions until they are needed, which will be after extension binding, when the type system is ready to go.

No tests because there is no way to reproduce this crash from a user module, and even then it’s not possible until later in this PR.