disable some assertions to quickly workaround 'try?' bug
If a 'try?' expression is implicitly coerced to 'Any',
assertions are tripped during SILGen, but we end up with
functionally-correct code emission. This is a
long-standing bug (since at least 5.0.3) in the compiler.
The best fix appears to be a minor overhaul of the SILGen
emission for all 'try?' expressions.

For now, this patch simply turns off the two assertions
that stand in the way of compiling such 'try?'
expressions for people using an asserts-enabled build
of the compiler.

A simple reproducer for this bug in an asserts build is:

    let _: Any = try? {() throws in 3 }()