[NFC] Allow many TypeRefinementContexts per decl
This PR modifies TypeRefinementContextBuilder to separate the implicit TRC around accessors from the creation of a new TRC for other declarations. This ability to introduce several TRCs for a single declaration will come in handy later.

It also refines the modeling of TRCs, particularly for properties and accessors, in a few others ways. Those changes are also currently NFC, but will become important down the line when function bodies start to become their own TRCs.