[Tests] Mark TypeRoundTrip as unsupported on use_os_stdlib.
When use_os_stdlib is set, we use the system Swift libraries, rather than
the ones we've just built.  Older Swift libraries don't have some of the
required functions to run this test, and since the test is supposed to be
testing the library rather than the compiler, it makes sense to just mark
it as unsupported in that configuration.