ClosureSpecializer: avoid an infinite optimization loop.
Avoid an infinite specialization loop caused by repeated runs of the ClosureSpecializer and CapturePropagation.
CapturePropagation propagates constant function-literals.
Such function specializations can then be optimized again by the ClosureSpecializer and so on.

This happens if a closure argument is called _and_ referenced in another closure, which is passed to a recursive call. E.g.

  func foo(_ c: @escaping () -> ()) {
    c()
    foo({ c() })
  }

rdar://80752327