Differentiation: correct the storage annotation for runtime functions
The runtime support functions are currently vended by
swiftCore.{dll,dylib,so} rather than
swift_Differentiation.{dll,dylib,so}.  This corrects the annotations to
indicate that reality.  This never could have worked as declared in the
first place.  swift_Differentiation never links against swiftRuntime
and swiftCore effectively whole-archives swiftRuntime into itself.  This
change is now reflecting that reality.  If the desire is to move this
(even on Darwin, where this may have already shipped and thus would
break ABI), we could split up the swiftRuntime into swiftRuntime and
swiftDifferentiationRuntime and link that as we do with the runtime into
swiftCore.