IRGen: convert an argument to deal with API changes
The current API in the associated branch point in LLVM has the following
signature for the `CoverageFiklenamesSectionWriter` ctor:

```c++
CoverageFilenamesSectionWriter(llvm::ArrayRef<llvm::StringRef>);
```

We currently had a `llvm::SmallVector<std::string, 8>` which could be
implicitly converted to `llvm::ArrayRef<std::string>`.  However, that is
not implicitly converted to `llvm::ArrayRef<llvm::StringRef>`.
Unfortunately, we currently have to convert `std::vector<StringRef>` to
`llvm::SmallVector<std::string, 8>` due to the file remapping, which
needs storage.  This storage then needs another conversion layer to get
the correct type setup.

The desired API would be available with
5fbd1a333aa1a0b70903d036b98ea56c51ae5224 (D95753), however that does not
cleanly apply and is a large refactoring.