[Preset] Add linux crosscompile wasm preset (#39183)
This preset will be used by community CI for WebAssembly to ensure [single-threaded-runtime mode won't be broken again](https://github.com/apple/swift/pull/39092#issuecomment-908992354) and enable us testing IRGen for WebAssembly in upstream.

Note that the current preset doesn't build stdlib for Wasm for now because some essential patches haven't been ported yet and build-script doesn't support it. But that should be enabled after that's supported.

This change is a part of SR-15168.