Disabling driver-time-compilation test
This test is pretty brittle on such a short file. If a member of the
timer output is zero, that member is omitted from the output. With how
short the file is, various fields come back zero occasionally.

On the Linux CI bot, there was a run with no "System Time", while I was
seeing it failing locally due to other missing time fields.